test: extract compiler tests + make #282 contract test cross-platform
Two CI failures on PR #283 needed fixing:

1. `Reject .rs files over 1000 LOC` -- compiler.rs was at 1064, over
   the 1000-LOC repo gate. Extracted the entire `mod tests { ... }`
   block to a sibling `compiler_tests.rs` and wired it back via
   `#[cfg(test)] #[path = "compiler_tests.rs"] mod tests;`. compiler.rs
   drops to 668 LOC; tests file is 418 LOC. No behavior change.

2. `Check (macos-latest)` and `(windows-latest)` failed on the new
   `compile_path_contract_pairs_cwd_and_output_arg_for_282` regression
   test. Cause: `compile_cwd_from_output` canonicalizes its result
   (resolving `/var` -> `/private/var` on macOS, and adding then
   stripping the `\?\` extended-length prefix on Windows). The test's
   `abs_output` was built from the raw `tempfile::tempdir()` path,
   which differs from the canonicalized form -- so the final equality
   assertion compared two spellings of the same physical file. Fixed
   by normalizing the tempdir base up front: canonicalize, then strip
   the Windows `\?\` prefix, so both sides of the assertion stay on
   the same form `compile_cwd_from_output` produces.

Local: `soldr cargo test -p fbuild-build --lib` -- 543 passed.

Refs #282.
